PiperOrigin-RevId: 235993145

+++ tensorflow/lite/toco/graph_transformations/group_bidirectional_sequence_ops.cc
@@ -135,7 +135,7 @@ bool CheckTwoUnidirectionalSequenceOpsAreValid(
     }
 
     // fw_lstm & bw_lstm should point to the same input, but reversed sequence.
-    for (int i = 0; i < first_fw_sequence_op_input->inputs.size(); ++i) {
+    for (size_t i = 0; i < first_fw_sequence_op_input->inputs.size(); ++i) {
       if (first_fw_sequence_op_input->inputs[i] !=
           first_bw_sequence_op_input
               ->inputs[first_fw_sequence_op_input->inputs.size() - i - 1]) {
@@ -296,7 +296,7 @@ void RewireBidirectionalSequenceSequenceOpsConnections(
   }
   string cur_fw_input = input_array_name;
   string cur_bw_input = input_array_name;
-  for (int i = 0; i < bidirectional_sequence_ops.size(); ++i) {
+  for (size_t i = 0; i < bidirectional_sequence_ops.size(); ++i) {
     DeleteArrayIfUsedOnce(bidirectional_sequence_ops[i]->inputs[0], model);
     bidirectional_sequence_ops[i]->inputs[0] = cur_fw_input;
     if (i != 0) {
@@ -329,7 +329,7 @@ void RewireFinalUnpackOutputs(const UnpackOperator& original_unpack_operator,
   (*final_unpack_operator)->axis = original_unpack_operator.axis;
   (*final_unpack_operator)->num = original_unpack_operator.num;
 
-  for (int i = 0; i < original_unpack_operator.outputs.size(); ++i) {
+  for (size_t i = 0; i < original_unpack_operator.outputs.size(); ++i) {
     const string& output_array_name = original_unpack_operator.outputs[i];
     const string& final_unpack_output_array_name = AvailableArrayName(
         *model, "bidirectional_sequence_unpack_" + std::to_string(i));
